% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_ancestor_type_map}
\alias{build_ancestor_type_map}
\title{Build Ancestor Map for Cell Ontology}
\usage{
build_ancestor_type_map(cl)
}
\arguments{
\item{cl}{An ontologyIndex object for Cell Ontology.}
}
\value{
Named list: clid → character vector of ancestor clids (including self)
}
\description{
For each CL ID, store all parent/ancestor CL IDs.
}
