% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{build_cl_term_map}
\alias{build_cl_term_map}
\title{Build Name/Synonym-to-CL-ID Map from Ontology Object}
\usage{
build_cl_term_map(cl, verbose = TRUE)
}
\arguments{
\item{cl}{An ontologyIndex CL ontology object.}

\item{verbose}{Logical; print mapping stats.}
}
\value{
Data.frame with columns: key, clid, cl_label.
}
\description{
Constructs a lookup table mapping CL names and synonyms to CL IDs.
}
\examples{
# cl <- ontologyIndex::get_ontology(
#   "http://purl.obolibrary.org/obo/cl.obo",
#   extract_tags = "everything"
# )
# cl_term_map <- build_cl_term_map(cl)
}
