% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_ontology_graph}
\alias{build_ontology_graph}
\title{Build igraph Representation of Ontology from Parent Relationships}
\usage{
build_ontology_graph(ontology)
}
\arguments{
\item{ontology}{An ontologyIndex object with a $parents list (e.g., Cell Ontology).}
}
\value{
An igraph object, nodes named by ontology term IDs.
}
\description{
Given an ontologyIndex object (e.g., as returned by \code{ontologyIndex::get_ontology()}),
constructs an igraph directed graph using 'is_a' parent relationships.
}
\examples{
# cl <- ontologyIndex::get_ontology(
#   "http://purl.obolibrary.org/obo/cl.obo",
#   extract_tags = "everything"
# )
# graph <- build_ontology_graph(cl)
}
