% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_mean_ontology_distance}
\alias{calculate_mean_ontology_distance}
\title{Calculate Mean Shortest Path Distance in CL Ontology}
\usage{
calculate_mean_ontology_distance(clid_df, graph, verbose = TRUE)
}
\arguments{
\item{clid_df}{Data.frame with columns clid1 and clid2 (CL IDs).}

\item{graph}{igraph object representing CL ontology DAG.}

\item{verbose}{Print progress and stats.}
}
\value{
List: mean_distance, inverse_mean, dist_vector.
}
\description{
For each row of a CL ID dataframe, calculates the ontology graph shortest path distance between pairs.
}
\examples{
# res <- calculate_mean_ontology_distance(clid_df, graph)
}
