% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcluster_anno.R
\name{combine_parent_subcluster_markers_internal}
\alias{combine_parent_subcluster_markers_internal}
\title{Combine Parent and Subcluster Markers (Internal)}
\usage{
combine_parent_subcluster_markers_internal(
  parent_markers,
  subcluster_markers,
  subset_seurat,
  subcluster_col,
  original_cluster_col,
  topgenenumber = 10
)
}
\description{
Combine Parent and Subcluster Markers (Internal)
}
\keyword{internal}
