% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.add_cl_names_to_combinations}
\alias{.add_cl_names_to_combinations}
\title{Add Cell Type Names to Unique Combinations}
\usage{
.add_cl_names_to_combinations(unique_combinations, cl_ontology)
}
\arguments{
\item{unique_combinations}{Data.frame with unique manual_clid/predicted_clid pairs}

\item{cl_ontology}{ontologyIndex CL object}
}
\value{
Data.frame with cell type names added for each unique combination
}
\description{
Add Cell Type Names to Unique Combinations
}
\keyword{internal}
