% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.classify_cl_matches}
\alias{.classify_cl_matches}
\title{Classify Unique CL ID Combinations (Internal)}
\usage{
.classify_cl_matches(unique_pairs, cl_ontology, graph, scoring_weights)
}
\arguments{
\item{unique_pairs}{Data.frame with manual_clid and predicted_clid columns}

\item{cl_ontology}{ontologyIndex CL object}

\item{graph}{igraph object}

\item{scoring_weights}{Named vector of match type weights}
}
\value{
Data.frame with match classifications
}
\description{
Classify Unique CL ID Combinations (Internal)
}
\keyword{internal}
