% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.get_direct_children}
\alias{.get_direct_children}
\title{Get Direct Children for CL IDs (Internal)}
\usage{
.get_direct_children(clids, cl_ontology)
}
\arguments{
\item{clids}{Character vector of CL IDs}

\item{cl_ontology}{ontologyIndex CL object}
}
\value{
Character vector of unique child CL IDs
}
\description{
Get Direct Children for CL IDs (Internal)
}
\keyword{internal}
