% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.get_node_depths}
\alias{.get_node_depths}
\title{Calculate Node Depths in Ontology Graph (Internal)}
\usage{
.get_node_depths(clids, graph)
}
\arguments{
\item{clids}{Character vector of CL IDs}

\item{graph}{igraph object}
}
\value{
Named numeric vector of depths from root
}
\description{
Calculate Node Depths in Ontology Graph (Internal)
}
\keyword{internal}
