% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.get_pairwise_distances}
\alias{.get_pairwise_distances}
\title{Calculate Pairwise Distances Between CL IDs (Internal)}
\usage{
.get_pairwise_distances(pairs, graph)
}
\arguments{
\item{pairs}{Data.frame with 'from' and 'to' columns containing CL IDs}

\item{graph}{igraph object}
}
\value{
Numeric vector of shortest path distances
}
\description{
Calculate Pairwise Distances Between CL IDs (Internal)
}
\keyword{internal}
