% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.print_results_summary}
\alias{.print_results_summary}
\title{Print Results Summary (Internal)}
\usage{
.print_results_summary(summary_stats)
}
\arguments{
\item{summary_stats}{List with summary statistics}
}
\description{
Print Results Summary (Internal)
}
\keyword{internal}
