% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.process_distance_one_relationships}
\alias{.process_distance_one_relationships}
\title{Process Distance-1 Relationships (Parent/Child) (Internal)}
\usage{
.process_distance_one_relationships(result_df, cl_ontology, scoring_weights)
}
\arguments{
\item{result_df}{Data.frame with classification results}

\item{cl_ontology}{ontologyIndex CL object}

\item{scoring_weights}{Named vector of scoring weights}
}
\value{
Updated result_df
}
\description{
Process Distance-1 Relationships (Parent/Child) (Internal)
}
\keyword{internal}
