% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.process_distance_two_relationships}
\alias{.process_distance_two_relationships}
\title{Process Distance-2 Relationships (Siblings) (Internal)}
\usage{
.process_distance_two_relationships(
  result_df,
  cl_ontology,
  graph,
  scoring_weights
)
}
\arguments{
\item{result_df}{Data.frame with classification results}

\item{cl_ontology}{ontologyIndex CL object}

\item{graph}{igraph object}

\item{scoring_weights}{Named vector of scoring weights}
}
\value{
Updated result_df
}
\description{
Process Distance-2 Relationships (Siblings) (Internal)
}
\keyword{internal}
