% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcluster_anno.R
\name{get_all_descendant_names}
\alias{get_all_descendant_names}
\title{Get All Descendant Names from an Ontology Term}
\usage{
get_all_descendant_names(ontology, term_name)
}
\arguments{
\item{ontology}{An ontology object from \pkg{ontologyIndex}.}

\item{term_name}{Character. Ontology term name to find descendants for.}
}
\value{
Character vector of descendant term names (unique).
}
\description{
Given a term name, returns all descendant term names (recursive) from a provided ontology.
Uses \code{get_descendants()} from the \pkg{ontologyIndex} package.
}
