% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_cl_direct_children}
\alias{get_cl_direct_children}
\title{Get Direct Children for CL IDs}
\usage{
get_cl_direct_children(clids, cl_ontology)
}
\arguments{
\item{clids}{Character vector of CL IDs (e.g., "CL:0000066")}

\item{cl_ontology}{ontologyIndex CL object from \code{ontologyIndex::get_ontology()}}
}
\value{
Character vector of unique child CL IDs
}
\description{
Retrieve direct child terms for given Cell Ontology IDs.
}
\examples{
\dontrun{
cl <- ontologyIndex::get_ontology("http://purl.obolibrary.org/obo/cl.obo")
children <- get_cl_direct_children("CL:0000066", cl)  # epithelial cell children
print(children)
}
}
