% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_cl_direct_parents}
\alias{get_cl_direct_parents}
\title{Get Direct Parents for CL IDs}
\usage{
get_cl_direct_parents(clids, cl_ontology)
}
\arguments{
\item{clids}{Character vector of CL IDs (e.g., "CL:0000066")}

\item{cl_ontology}{ontologyIndex CL object from \code{ontologyIndex::get_ontology()}}
}
\value{
Character vector of unique parent CL IDs
}
\description{
Retrieve direct parent terms for given Cell Ontology IDs.
}
\examples{
\dontrun{
cl <- ontologyIndex::get_ontology("http://purl.obolibrary.org/obo/cl.obo")
parents <- get_cl_direct_parents("CL:0000066", cl)  # epithelial cell parents
print(parents) # CL:0000255 eukaryotic cell
}
}
