% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_cl_names}
\alias{get_cl_names}
\title{Get Cell Type Names for CL IDs}
\usage{
get_cl_names(clids, cl_ontology)
}
\arguments{
\item{clids}{Character vector of CL IDs}

\item{cl_ontology}{ontologyIndex CL object from \code{ontologyIndex::get_ontology()}}
}
\value{
Named character vector of cell type names
}
\description{
Retrieve official cell type names from Cell Ontology for given CL IDs.
}
\examples{
\dontrun{
cl <- ontologyIndex::get_ontology("http://purl.obolibrary.org/obo/cl.obo")
names <- get_cl_names(c("CL:0000066", "CL:0000057"), cl)
print(names)
}
}
