% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper Operator for Default Values (Internal)}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Default value if x is NULL/NA/empty}
}
\value{
x if valid, otherwise y
}
\description{
Helper Operator for Default Values (Internal)
}
\keyword{internal}
