% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_annotation_distance_ontology}
\alias{score_annotation_distance_ontology}
\title{Score Annotation Distance Based on Cell Ontology}
\usage{
score_annotation_distance_ontology(
  seurat_obj,
  col1,
  col2,
  cl_term_map = GPTAnno::cl_term_map,
  graph,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{col1}{Character. Metadata column name for reference/manual annotation.}

\item{col2}{Character. Metadata column for predicted/auto annotation.}

\item{cl_term_map}{Data.frame; defaults to package built-in map, or returned by \code{build_cl_term_map()}.}

\item{graph}{Ontology graph built from CL object.}

\item{verbose}{Print mapping/progress.}
}
\value{
List with mean distance and mapping dataframe result.
}
\description{
Computes the mean shortest-path distance between annotations in two columns using CL ontology.
(Distance-based complement to score_annotation_agreement_ontology.)
}
\examples{
\dontrun{
# Build ontology and graph (one-time setup)
cl <- ontologyIndex::get_ontology("http://purl.obolibrary.org/obo/cl.obo",
 extract_tags = "everything")
graph <- build_ontology_graph(cl)
cl_term_map <- build_cl_term_map(cl) # if need the up-to-date mapping
# Score annotation distance
dist_res <- score_annotation_distance_ontology(
  seurat_obj, col1 = "celltype_manual", col2 = "celltype_parent",
  cl_term_map = cl_term_map, # leave it blank will use the built-in mapping
  graph = ontology_graph
)
print(dist_res$mean_distance)
}

}
