% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcluster_anno.R
\name{assign_inherited_subcluster_annotations}
\alias{assign_inherited_subcluster_annotations}
\title{Assign Inherited Subcluster Annotations}
\usage{
assign_inherited_subcluster_annotations(
  full_seurat,
  annotation_sub_inherited,
  subcluster_annotation_col_prefix = "annotated_sub_",
  final_colname = "celltype_subcluster",
  parent_colname = "celltype_parent",
  clean_labels = TRUE
)
}
\arguments{
\item{full_seurat}{Seurat object for full dataset.}

\item{annotation_sub_inherited}{Nested list of inherited subcluster annotation objects.}

\item{subcluster_annotation_col_prefix}{Prefix for subcluster annotation columns.}

\item{final_colname}{Name of new metadata column.}

\item{parent_colname}{Column for parent annotation.}

\item{clean_labels}{Whether to clean annotation labels of " (not in GO)" suffix.}
}
\value{
The updated Seurat object with inherited subcluster annotation column.
}
\description{
Combines subcluster and parent annotations to create a new celltype_subcluster column.
}
