% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPTCelltyp_mapping}
\alias{GPTCelltyp_mapping}
\title{Reference Mapping: GPTCelltyp_mapping}
\format{
A named character vector (or data.frame if you want). Names are GPT-4 annotations, values are CL names.
}
\source{
\url{https://doi.org/10.1038/s41592-024-02235-4}
}
\usage{
data(GPTCelltyp_mapping)
}
\description{
A curated data frame mapping GPT-4-predicted cell type annotations to standardized Cell Ontology (CL) names,
based on the supplementary material from paper: Hou, Wenpin, and Zhicheng Ji. "Assessing GPT-4 for cell type annotation in single-cell RNA-seq analysis." Nature methods 21.8 (2024): 1462-1465.
}
\examples{
data(GPTCelltyp_mapping)
head(GPTCelltyp_mapping)
}
\seealso{
\code{\link{clean_and_match_annotation}}, \code{\link{map_celltypes_to_cl}}
}
\keyword{datasets}
