% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPTCelltype_mapping}
\alias{GPTCelltype_mapping}
\title{GPTCelltype_mapping: Reference Mapping of GPT-4 Annotations to Cell Ontology}
\format{
A named character vector. Names are GPT-4 annotation results, values are CL names.
}
\source{
\url{https://doi.org/10.1038/s41592-024-02235-4}
}
\usage{
data(GPTCelltype_mapping)
}
\description{
A curated mapping from GPT-4-predicted cell type names to standardized Cell Ontology (CL) terms,
based on the supplementary material from:
Hou, Wenpin, and Zhicheng Ji. "Assessing GPT-4 for cell type annotation in single-cell RNA-seq analysis." \emph{Nature Methods} 21.8 (2024): 1462-1465.
}
\examples{
data(GPTCelltype_mapping)
head(GPTCelltype_mapping)
}
\seealso{
\code{\link{clean_and_match_annotation}}, \code{\link{map_celltypes_to_cl}}
}
